backend/requirements.txt
fastapi
uvicorn
gunicorn
python-multipart
opencv-python-headless
numpy
tensorflow-cpu
Pillow
google-generativeai
pydantic
pydantic-settings
slowapi
